Reestructure of common and app modules
Moved components into components folder in order to be more organized
Moved services into its own folder for organization purposes
Updated all modules to have valid dependencies
