+++ src/app/app.module.ts
@@ -1,13 +1,13 @@
 import { BrowserModule } from '@angular/platform-browser';
-import { CommonComponentsModule } from 'common-components/common-components.module';
+import { CommonModule } from '@common/common.module';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 
-import { AppComponent } from 'app/app.component';
-import { MainViewComponent } from 'app/main-view.component';
-import { MoviesListComponent } from 'app/movies-list.component';
-import { MovieSummaryComponent } from 'app/movie-summary.component';
+import { AppComponent } from '@app/components/app.component';
+import { MainViewComponent } from '@app/components/main-view.component';
+import { MoviesListComponent } from '@app/components/movies-list.component';
+import { MovieSummaryComponent } from '@app/components/movie-summary.component';
 
 @NgModule({
     declarations: [
@@ -18,7 +18,7 @@ import { MovieSummaryComponent } from 'app/movie-summary.component';
     ],
     imports: [
         BrowserModule,
-        CommonComponentsModule,
+        CommonModule,
         FormsModule,
         HttpModule
     ],
