+++ src/app/components/app.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-root',
+    templateUrl: 'app.component.html',
+    styleUrls: [
+        'app.component.scss'
+    ]
+})
+export class AppComponent {
+    title = 'app works!';
+}
