+++ src/app/components/main-view.component.scss
