+++ src/app/components/movies-list.component.scss
