--- src/app/main-view.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'main-view',
-    templateUrl: 'main-view.component.html',
-    styleUrls: [
-        'main-view.component.scss'
-    ]
-})
-export class MainViewComponent {
-
-}
