--- src/common-components/common-components.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-
-import { HeaderComponent } from 'common-components/header.component';
-
-@NgModule({
-    declarations: [
-        HeaderComponent
-    ],
-    exports: [
-        HeaderComponent
-    ],
-    imports: [
-
-    ]
-})
-export class CommonComponentsModule { }
