+++ src/globals.scss
@@ -1,5 +1,5 @@
 $black: #000000;
-$blue: #164398;
+$blue: #2653A8;
 $green: #15D86C;
 $light-black: #24292D;
 $light-blue: #3186C5;
