+++ src/styles.scss
@@ -11,4 +11,5 @@
 
 * {
     box-sizing: border-box;
+    font-family: 'Open Sans', arial, helvetica, sans-serif;
 }
